chore: deprecates getPayloadHMR in favor of simpler getPayload
