+++ docs/live-preview/server.mdx
@@ -34,11 +34,11 @@ Then, render the `RefreshRouteOnSave` component anywhere in your `page.tsx`. Her
 
 ```tsx
 import { RefreshRouteOnSave } from './RefreshRouteOnSave.tsx'
-import { getPayloadHMR } from '@payloadcms/next/utilities'
+import { getPayload } from 'payload'
 import config from '../payload.config'
 
 export default async function Page() {
-  const payload = await getPayloadHMR({ config })
+  const payload = await getPayload({ config })
 
   const page = await payload.findByID({
     collection: 'pages',
