+++ examples/live-preview/payload/README.md
@@ -107,11 +107,11 @@ Then, render `RefreshRouteOnSave` anywhere in your `page.tsx`. Here's an example
 
 ```tsx
 import { RefreshRouteOnSave } from './RefreshRouteOnSave.tsx'
-import { getPayloadHMR } from '@payloadcms/next/utilities'
+import { getPayload } from 'payload'
 import config from '../payload.config'
 
 export default async function Page() {
-  const payload = await getPayloadHMR({ config })
+  const payload = await getPayload({ config })
 
   const page = await payload.find({
     collection: 'pages',
