+++ packages/next/src/exports/utilities.ts
@@ -3,6 +3,6 @@ export { addDataAndFileToRequest } from '../utilities/addDataAndFileToRequest.js
 export { addLocalesToRequestFromData, sanitizeLocales } from '../utilities/addLocalesToRequest.js'
 export { createPayloadRequest } from '../utilities/createPayloadRequest.js'
 export { getNextRequestI18n } from '../utilities/getNextRequestI18n.js'
-export { getPayloadHMR, reload } from '../utilities/getPayloadHMR.js'
+export { getPayloadHMR } from '../utilities/getPayloadHMR.js'
 export { headersWithCors } from '../utilities/headersWithCors.js'
 export { mergeHeaders } from '../utilities/mergeHeaders.js'
