+++ packages/next/src/layouts/Root/index.tsx
@@ -5,12 +5,11 @@ import { rtlLanguages } from '@payloadcms/translations'
 import { RootProvider } from '@payloadcms/ui'
 import '@payloadcms/ui/scss/app.scss'
 import { headers as getHeaders, cookies as nextCookies } from 'next/headers.js'
-import { checkDependencies, parseCookies } from 'payload'
+import { checkDependencies, getPayload, parseCookies } from 'payload'
 import React from 'react'
 
 import { getNavPrefs } from '../../elements/Nav/getNavPrefs.js'
 import { getClientConfig } from '../../utilities/getClientConfig.js'
-import { getPayloadHMR } from '../../utilities/getPayloadHMR.js'
 import { getRequestLanguage } from '../../utilities/getRequestLanguage.js'
 import { getRequestTheme } from '../../utilities/getRequestTheme.js'
 import { initReq } from '../../utilities/initReq.js'
@@ -100,7 +99,7 @@ export const RootLayout = async ({
     headers,
   })
 
-  const payload = await getPayloadHMR({ config })
+  const payload = await getPayload({ config })
 
   const { i18n, permissions, req, user } = await initReq(config)
