+++ packages/payload/src/index.ts
@@ -6,6 +6,7 @@ import { spawn } from 'child_process'
 import crypto from 'crypto'
 import { fileURLToPath } from 'node:url'
 import path from 'path'
+import WebSocket from 'ws'
 
 import type { AuthArgs } from './auth/operations/auth.js'
 import type { Result as ForgotPasswordResult } from './auth/operations/forgotPassword.js'
@@ -17,7 +18,6 @@ import type { Options as VerifyEmailOptions } from './auth/operations/local/veri
 import type { Result as LoginResult } from './auth/operations/login.js'
 import type { Result as ResetPasswordResult } from './auth/operations/resetPassword.js'
 import type { AuthStrategy, User } from './auth/types.js'
-import type { ImportMap } from './bin/generateImportMap/index.js'
 import type {
   BulkOperationResult,
   Collection,
@@ -25,6 +25,8 @@ import type {
   SelectFromCollectionSlug,
   TypeWithID,
 } from './collections/config/types.js'
+
+import { generateImportMap, type ImportMap } from './bin/generateImportMap/index.js'
 export type { FieldState } from './admin/forms/Form.js'
 export type * from './admin/types.js'
 import type { NonNever } from 'ts-essentials'
@@ -726,30 +728,137 @@ const initialized = new BasePayload()
 
 export default initialized
 
-let cached = global._payload
+let cached: {
+  payload: null | Payload
+  promise: null | Promise<Payload>
+  reload: boolean | Promise<void>
+  ws: null | WebSocket
+} = global._payload
 
 if (!cached) {
-  cached = global._payload = { payload: null, promise: null }
+  cached = global._payload = { payload: null, promise: null, reload: false, ws: null }
 }
 
-export const getPayload = async (options: InitOptions): Promise<BasePayload> => {
+export const reload = async (
+  config: SanitizedConfig,
+  payload: Payload,
+  skipImportMapGeneration?: boolean,
+): Promise<void> => {
+  if (typeof payload.db.destroy === 'function') {
+    await payload.db.destroy()
+  }
+
+  payload.config = config
+
+  payload.collections = config.collections.reduce((collections, collection) => {
+    collections[collection.slug] = {
+      config: collection,
+      customIDType: payload.collections[collection.slug]?.customIDType,
+    }
+    return collections
+  }, {})
+
+  payload.globals = {
+    config: config.globals,
+  }
+
+  // TODO: support HMR for other props in the future (see payload/src/index init()) that may change on Payload singleton
+
+  // Generate types
+  if (config.typescript.autoGenerate !== false) {
+    // We cannot run it directly here, as generate-types imports json-schema-to-typescript, which breaks on turbopack.
+    // see: https://github.com/vercel/next.js/issues/66723
+    void payload.bin({
+      args: ['generate:types'],
+      log: false,
+    })
+  }
+
+  // Generate component map
+  if (skipImportMapGeneration !== true && config.admin?.importMap?.autoGenerate !== false) {
+    await generateImportMap(config, {
+      log: true,
+    })
+  }
+
+  await payload.db.init()
+  if (payload.db.connect) {
+    await payload.db.connect({ hotReload: true })
+  }
+}
+
+export const getPayload = async (
+  options: Pick<InitOptions, 'config' | 'importMap'>,
+): Promise<Payload> => {
   if (!options?.config) {
     throw new Error('Error: the payload config is required for getPayload to work.')
   }
 
   if (cached.payload) {
+    if (cached.reload === true) {
+      let resolve: () => void
+
+      // getPayload is called multiple times, in parallel. However, we only want to run `await reload` once. By immediately setting cached.reload to a promise,
+      // we can ensure that all subsequent calls will wait for the first reload to finish. So if we set it here, the 2nd call of getPayload
+      // will reach `if (cached.reload instanceof Promise) {` which then waits for the first reload to finish.
+      cached.reload = new Promise((res) => (resolve = res))
+      const config = await options.config
+      await reload(config, cached.payload)
+
+      resolve()
+    }
+
+    if (cached.reload instanceof Promise) {
+      await cached.reload
+    }
+
+    if (options?.importMap) {
+      cached.payload.importMap = options.importMap
+    }
     return cached.payload
   }
 
+  // eslint-disable-next-line @typescript-eslint/no-misused-promises
   if (!cached.promise) {
+    // no need to await options.config here, as it's already awaited in the BasePayload.init
     cached.promise = new BasePayload().init(options)
   }
 
   try {
     cached.payload = await cached.promise
+
+    if (
+      !cached.ws &&
+      process.env.NODE_ENV !== 'production' &&
+      process.env.NODE_ENV !== 'test' &&
+      process.env.DISABLE_PAYLOAD_HMR !== 'true'
+    ) {
+      try {
+        const port = process.env.PORT || '3000'
+        cached.ws = new WebSocket(
+          `ws://localhost:${port}${process.env.NEXT_BASE_PATH ?? ''}/_next/webpack-hmr`,
+        )
+
+        cached.ws.onmessage = (event) => {
+          if (typeof event.data === 'string') {
+            const data = JSON.parse(event.data)
+
+            if ('action' in data && data.action === 'serverComponentChanges') {
+              cached.reload = true
+            }
+          }
+        }
+      } catch (_) {
+        // swallow e
+      }
+    }
   } catch (e) {
     cached.promise = null
     throw e
+  }
+
+  if (options?.importMap) {
+    cached.payload.importMap = options.importMap
   }
 
   return cached.payload
